pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatgpt-cli"
version = "0.4.5"
authors = [
    { name = "Marco Lardera", email = "larderamarco@hotmail.com" },
    { name = "Tadeas Fort", email = "business@tadeasfort.com" }
]
description = "Simple yet effective command line client for chatting with ChatGPT using the official API."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "importlib-metadata==8.2.0",
    "litellm>=1.44.5",  # Updated version
    "prompt-toolkit==3.0.47",
    "pydantic-core==2.20.1",
    "pyyaml>=6.0.2",
    "rich==13.7.1",
    "typing-extensions==4.12.2",
    "xdg-base-dirs==6.0.1",
    "hydra-core>=1.3.2",
    "pydantic>=2.8.2",
    "pydantic-collections>=0.6.0",
    "fsspec>=2024.6.1",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.8.0",
]
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pdm]
[project.scripts]
chatgpt-cli = "chatgpt_cli.main:main"

[project.urls]
"Homepage" = "https://github.com/marcolardera/chatgpt-cli"
"Bug Tracker" = "https://github.com/marcolardera/chatgpt-cli/issues"

[tool.setuptools.package-data]
"chatgpt_cli" = ["config_template.yaml"]

[tool.ruff]
line-length = 120
